% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonzero.R
\name{nonzero}
\alias{nonzero}
\title{Return selected variables}
\usage{
nonzero(object, ...)
}
\arguments{
\item{object}{a model object for which the set of selected variables is desired}

\item{...}{additional arguments, e.g. tuning parameters}
}
\value{
The form of the value returned by \code{nonzero} depends on the class of its
argument. See the documentation of the particular methods for details of what
is produced by that method.
}
\description{
\code{nonzero} is a generic function for returning the set of variables selected by a model
}
