% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonzero.singleton.R
\name{nonzero.singleton}
\alias{nonzero.singleton}
\title{Return selected variables from a \code{singleton} object}
\usage{
\method{nonzero}{singleton}(object, ...)
}
\arguments{
\item{object}{an object of class \code{singleton}}

\item{...}{ignored}
}
\description{
Returns NULL. Intended for internal use only
}
