% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.singleton.R
\name{predict.singleton}
\alias{predict.singleton}
\title{Make predictions from a ``singleton'' object}
\usage{
\method{predict}{singleton}(object, type = c("response", "class", "nonzero"), ...)
}
\arguments{
\item{object}{an object of class \code{singleton}}

\item{type}{type of prediction to be returned, "response" or "class"}

\item{...}{ignored}
}
\value{
The value of the singleton
}
\description{
Returns the value stored in the singleton. Intended for internal use only.
}
